back-end/requirements.txt
fastapi
uvicorn[standard]
PyPDF2
fpdf
reportlab
python-multipart
pydantic
python-docx
